workaround (duplicate) types for now
