+++ src/types/index.ts
@@ -1,4 +1,3 @@
-export * from "./error";
 export * from "./fetch";
 export * from "./global";
 export * from "./h3";
